refactor schedulers, updated WebRTCClientTest imports
