+++ syftlib/src/main/java/org/openmined/syft/interfaces/SignallingInterface.kt
@@ -1,20 +1,6 @@
-package org.openmined.syft
+package org.openmined.syft.interfaces
 
-import io.reactivex.Scheduler
-
-interface SignallingInterface {
-
-    /**
-     * computeThreadScheduler defines the thread on which observable runs
-     * @sample computeThreadScheduler Schedulers.io()
-     */
-    val computeThreadScheduler : Scheduler
-
-    /**
-     * calleeThreadScheduler defines the thread on which the callback to observable is run
-     * @sample calleeThreadScheduler AndroidSchedulers.MainThread()
-     */
-    val calleeThreadScheduler : Scheduler
+interface SignallingInterface : ProcessSchedulers {
 
     /**
      * Called in the UI activity as a response to message received from PyGrid
